pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "azure-storage-blob>=12.27.0",
    "flask>=3.1.2",
    "google-genai>=1.46.0",
    "openai>=2.6.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "werkzeug>=3.1.3",
]
